Thay Doi Double

+++ Form1.cs
@@ -18,7 +18,7 @@ namespace Caldemo
         private void btnCong_Click(object sender, EventArgs e)
         {
             double a = double.Parse(textBox1.Text);
-            double b = double.Parse(textBox2.Text);
+            int b = int.Parse(textBox2.Text);
             double c = a + b;
             label1.Text = c.ToString();
         }
